Fix resize for CAPI clusters
Clusters created by the magnum-capi-helm driver do not have a heat stack, so
resize currently fails as it depends on the heat stack for a list of worker
nodes. To fix this, we separate out worker count and list.

Worker count is used to validate resizing up and down. Use
cluster.node_count for this.

Worker list is used to select worker to be removed. This is populated by
heat, or set to [] if it is using capi.

Change-Id: I0d66fc02a7f2c608a4a0b09b98c343017e04ed41